Revert "Refactor App structure and update dependencies"
