+++ README.md
@@ -1,122 +1,1 @@
-# Haven Frontend
-
-A React Native frontend application with TypeScript, built with Expo.
-
-## 📁 Project Structure
-
-This project follows a professional folder structure designed for scalability and maintainability. For detailed information about the project organization, architecture patterns, and development guidelines, please refer to:
-
-**[📋 PROJECT_STRUCTURE.md](./PROJECT_STRUCTURE.md)**
-
-## 🚀 Getting Started
-
-### Prerequisites
-
-- Node.js (v16 or higher)
-- npm or yarn
-- Expo CLI
-- React Native development environment
-
-### Installation
-
-1. **Clone the repository:**
-   ```bash
-   git clone <repository-url>
-   cd Haven-frontend
-   ```
-
-2. **Install dependencies:**
-   ```bash
-   npm install
-   # or
-   yarn install
-   ```
-
-3. **Install Expo CLI globally (if not already installed):**
-   ```bash
-   npm install -g @expo/cli
-   ```
-
-### Running the Project
-
-1. **Start the development server:**
-   ```bash
-   npm start
-   # or
-   yarn start
-   # or
-   expo start
-   ```
-
-2. **Run on specific platforms:**
-   ```bash
-   # iOS simulator
-   npm run ios
-   # or
-   yarn ios
-
-   # Android emulator
-   npm run android
-   # or
-   yarn android
-
-   # Web browser
-   npm run web
-   # or
-   yarn web
-   ```
-
-### Development Commands
-
-```bash
-# Start development server
-npm start
-
-# Run on iOS
-npm run ios
-
-# Run on Android
-npm run android
-
-# Run on Web
-npm run web
-
-# TypeScript check
-npx tsc --noEmit
-
-# Clear cache
-expo start --clear
-```
-
-## 📱 Development Guidelines
-
-- Follow the project structure outlined in [PROJECT_STRUCTURE.md](./PROJECT_STRUCTURE.md)
-- Use TypeScript for type safety
-- Follow React Native best practices
-- Implement proper error handling
-- Write clean, maintainable code
-
-## 🔧 Tech Stack
-
-- **React Native** - Mobile app framework
-- **Expo** - Development platform
-- **TypeScript** - Type safety
-- **React Navigation** - Navigation
-- **AsyncStorage** - Local storage
-- **Context API** - State management
-
-## 📚 Additional Resources
-
-- [React Native Documentation](https://reactnative.dev/)
-- [Expo Documentation](https://docs.expo.dev/)
-- [TypeScript Documentation](https://www.typescriptlang.org/)
-- [React Navigation](https://reactnavigation.org/)
-
-## 🤝 Contributing
-
-1. Follow the project structure and coding standards
-2. Create feature branches for new functionality
-3. Write clear commit messages
-4. Test your changes thoroughly
-
-For detailed development guidelines and project architecture, see [PROJECT_STRUCTURE.md](./PROJECT_STRUCTURE.md).
+# Haven---frontend
