--- src/App.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { StatusBar } from 'expo-status-bar';
-import { AuthProvider } from './contexts';
-import AppNavigator from './navigation/AppNavigator';
-
-export default function App() {
-  return (
-    <AuthProvider>
-      <AppNavigator />
-      <StatusBar style="auto" />
-    </AuthProvider>
-  );
-}
