--- src/components/Button/Button.tsx
@@ -1,116 +1,0 @@
-import React from 'react';
-import { TouchableOpacity, Text, StyleSheet, ActivityIndicator, ViewStyle, TextStyle } from 'react-native';
-import theme from '../../styles/theme';
-
-interface ButtonProps {
-  title: string;
-  onPress: () => void;
-  variant?: 'primary' | 'secondary' | 'outline';
-  disabled?: boolean;
-  loading?: boolean;
-  fullWidth?: boolean;
-}
-
-const Button: React.FC<ButtonProps> = ({
-  title,
-  onPress,
-  variant = 'primary',
-  disabled = false,
-  loading = false,
-  fullWidth = false,
-}) => {
-  const getButtonStyle = (): ViewStyle[] => {
-    const baseStyle: ViewStyle[] = [styles.button];
-    
-    if (fullWidth) baseStyle.push(styles.fullWidth);
-    if (disabled) baseStyle.push(styles.disabled);
-    
-    switch (variant) {
-      case 'secondary':
-        baseStyle.push(styles.secondary);
-        break;
-      case 'outline':
-        baseStyle.push(styles.outline);
-        break;
-      default:
-        baseStyle.push(styles.primary);
-    }
-    
-    return baseStyle;
-  };
-
-  const getTextStyle = (): TextStyle[] => {
-    const baseStyle: TextStyle[] = [styles.text];
-    
-    switch (variant) {
-      case 'secondary':
-        baseStyle.push(styles.secondaryText);
-        break;
-      case 'outline':
-        baseStyle.push(styles.outlineText);
-        break;
-      default:
-        baseStyle.push(styles.primaryText);
-    }
-    
-    return baseStyle;
-  };
-
-  return (
-    <TouchableOpacity
-      style={getButtonStyle()}
-      onPress={onPress}
-      disabled={disabled || loading}
-      activeOpacity={0.8}
-    >
-      {loading ? (
-        <ActivityIndicator color={variant === 'outline' ? theme.colors.primary : 'white'} />
-      ) : (
-        <Text style={getTextStyle()}>{title}</Text>
-      )}
-    </TouchableOpacity>
-  );
-};
-
-const styles = StyleSheet.create({
-  button: {
-    paddingHorizontal: 20,
-    paddingVertical: 12,
-    borderRadius: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-    minHeight: 48,
-  },
-  fullWidth: {
-    width: '100%',
-  },
-  primary: {
-    backgroundColor: theme.colors.primary,
-  },
-  secondary: {
-    backgroundColor: theme.colors.secondary,
-  },
-  outline: {
-    backgroundColor: 'transparent',
-    borderWidth: 1,
-    borderColor: theme.colors.primary,
-  },
-  disabled: {
-    opacity: 0.6,
-  },
-  text: {
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  primaryText: {
-    color: 'white',
-  },
-  secondaryText: {
-    color: 'white',
-  },
-  outlineText: {
-    color: theme.colors.primary,
-  },
-});
-
-export default Button;
