--- src/components/index.ts
@@ -1,6 +1,0 @@
-// Centralized component exports
-export { default as Button } from './Button/Button';
-// Add more components here as they are created
-// export { default as Input } from './Input';
-// export { default as Card } from './Card';
-// export { default as Modal } from './Modal';
