--- src/contexts/index.ts
@@ -1,5 +1,0 @@
-// Centralized context exports
-export { AuthProvider, useAuthContext } from './AuthContext';
-// Add more contexts here as they are created
-// export { ThemeProvider, useThemeContext } from './ThemeContext';
-// export { SettingsProvider, useSettingsContext } from './SettingsContext';
