--- src/hooks/index.ts
@@ -1,6 +1,0 @@
-// Centralized hooks exports
-export { useAuth } from './useAuth';
-// Add more hooks here as they are created
-// export { useTheme } from './useTheme';
-// export { useNetwork } from './useNetwork';
-// export { useKeyboard } from './useKeyboard';
