--- src/screens/HomeScreen/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HomeScreen';
