--- src/screens/LoginScreen/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LoginScreen';
