--- src/screens/RegisterScreen/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RegisterScreen';
