--- src/screens/index.ts
@@ -1,7 +1,0 @@
-// Centralized screen exports
-export { default as HomeScreen } from './HomeScreen';
-// Add more screens here as they are created
-// export { default as ProfileScreen } from './ProfileScreen';
-// export { default as SettingsScreen } from './SettingsScreen';
-export { default as LoginScreen } from './LoginScreen'; 
-export { default as RegisterScreen } from './RegisterScreen';
