--- src/services/api.ts
@@ -1,85 +1,0 @@
-import { APP_CONFIG, API_ENDPOINTS } from '../constants';
-import { ApiResponse, User } from '../types';
-
-class ApiService {
-  private baseURL: string;
-  private timeout: number;
-
-  constructor() {
-    this.baseURL = APP_CONFIG.API_BASE_URL;
-    this.timeout = APP_CONFIG.TIMEOUT;
-  }
-
-  private async request<T>(
-    endpoint: string,
-    options: RequestInit = {}
-  ): Promise<ApiResponse<T>> {
-    const url = `${this.baseURL}${endpoint}`;
-    
-    const config: RequestInit = {
-      ...options,
-      headers: {
-        'Content-Type': 'application/json',
-        ...options.headers,
-      },
-    };
-
-    try {
-      const response = await fetch(url, config);
-      const data = await response.json();
-
-      if (!response.ok) {
-        throw new Error(data.message || 'Request failed');
-      }
-
-      return {
-        data,
-        status: 'success',
-        message: data.message,
-      };
-    } catch (error) {
-      return {
-        data: null as T,
-        status: 'error',
-        message: error instanceof Error ? error.message : 'Unknown error',
-      };
-    }
-  }
-
-  // Auth methods
-  async login(email: string, password: string): Promise<ApiResponse<{ token: string; user: User }>> {
-    return this.request(API_ENDPOINTS.AUTH.LOGIN, {
-      method: 'POST',
-      body: JSON.stringify({ email, password }),
-    });
-  }
-
-  async register(userData: Omit<User, 'id'>): Promise<ApiResponse<{ token: string; user: User }>> {
-    return this.request(API_ENDPOINTS.AUTH.REGISTER, {
-      method: 'POST',
-      body: JSON.stringify(userData),
-    });
-  }
-
-  // User methods
-  async getProfile(token: string): Promise<ApiResponse<User>> {
-    return this.request(API_ENDPOINTS.USER.PROFILE, {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-  }
-
-  async updateProfile(token: string, userData: Partial<User>): Promise<ApiResponse<User>> {
-    return this.request(API_ENDPOINTS.USER.UPDATE, {
-      method: 'PUT',
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-      body: JSON.stringify(userData),
-    });
-  }
-}
-
-export default new ApiService();
