--- src/services/index.ts
@@ -1,6 +1,0 @@
-// Centralized service exports
-export { default as ApiService } from './api';
-export { default as StorageService } from './storage';
-// Add more services here as they are created
-// export { default as NotificationService } from './notification';
-// export { default as AnalyticsService } from './analytics';
