--- src/types/auth/index.ts
@@ -1,62 +1,0 @@
-// Authentication and User types
-
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  avatar?: string;
-  role?: UserRole;
-  createdAt?: string;
-  updatedAt?: string;
-  isActive?: boolean;
-}
-
-export enum UserRole {
-  ADMIN = 'admin',
-  USER = 'user',
-  MODERATOR = 'moderator',
-}
-
-export interface AuthState {
-  user: User | null;
-  token: string | null;
-  refreshToken: string | null;
-  isAuthenticated: boolean;
-  loading: boolean;
-  error: string | null;
-}
-
-export interface AuthActions {
-  login: (email: string, password: string) => Promise<boolean>;
-  register: (userData: RegisterUserData) => Promise<boolean>;
-  logout: () => Promise<void>;
-  refreshToken: () => Promise<boolean>;
-  clearError: () => void;
-  updateProfile: (userData: Partial<User>) => Promise<boolean>;
-}
-
-export interface RegisterUserData {
-  name: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-}
-
-export interface LoginCredentials {
-  email: string;
-  password: string;
-}
-
-export interface AuthContextValue extends AuthState, AuthActions {}
-
-// Permission types
-export interface Permission {
-  id: string;
-  name: string;
-  description: string;
-}
-
-export interface UserPermissions {
-  userId: string;
-  permissions: Permission[];
-}
