--- src/utils/validation.ts
@@ -1,50 +1,0 @@
-import { VALIDATION_RULES } from '../constants';
-
-export const validateEmail = (email: string): boolean => {
-  return VALIDATION_RULES.EMAIL_REGEX.test(email);
-};
-
-export const validatePassword = (password: string): boolean => {
-  return password.length >= VALIDATION_RULES.PASSWORD_MIN_LENGTH;
-};
-
-export const validateName = (name: string): boolean => {
-  return name.trim().length >= VALIDATION_RULES.NAME_MIN_LENGTH;
-};
-
-export const validateRequired = (value: string): boolean => {
-  return value.trim().length > 0;
-};
-
-export const formatPhoneNumber = (phone: string): string => {
-  const cleaned = phone.replace(/\D/g, '');
-  const match = cleaned.match(/^(\d{3})(\d{3})(\d{4})$/);
-  if (match) {
-    return `(${match[1]}) ${match[2]}-${match[3]}`;
-  }
-  return phone;
-};
-
-export const formatCurrency = (amount: number): string => {
-  return new Intl.NumberFormat('en-US', {
-    style: 'currency',
-    currency: 'USD',
-  }).format(amount);
-};
-
-export const truncateText = (text: string, maxLength: number): string => {
-  if (text.length <= maxLength) return text;
-  return text.substring(0, maxLength) + '...';
-};
-
-export const debounce = <T extends (...args: any[]) => void>(
-  func: T,
-  delay: number
-): ((...args: Parameters<T>) => void) => {
-  let timeoutId: NodeJS.Timeout;
-  
-  return (...args: Parameters<T>) => {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => func(...args), delay);
-  };
-};
